refactor: GeminiEmbedding.ts -> embed.ts
