+++ packages/providers/google/src/embed.ts
