+++ packages/providers/google/src/index.ts
@@ -10,4 +10,4 @@ export * from "./types";
 export * from "./utils";
 export * from "./vertex";
 
-export * from "./GeminiEmbedding";
+export * from "./embed";
